Add back mod_vcard_ldap and change eldap stuff to work with binaries.
The octet strings are converted to octet lists by asn1rt. Not sure it's worth spending
more time on trying to keep it as binaries, so just doing list_to_binary/1 on them
before returning results from eldap.erl. If eldap is slow on heavy load, this would
be something to look at.